COMS - change unit test directory names for client and service tests
