+++ src/backend/TrafficCourts/TrafficCourts.Coms.Client.Test/ClientTests/AddPermissions.cs
@@ -1,0 +1,5 @@
+﻿namespace TrafficCourts.Coms.Client.Test.ClientTests;
+
+public class AddPermissions : ObjectManagementBase
+{
+}
