+++ src/backend/TrafficCourts/TrafficCourts.Coms.Client.Test/ClientTests/AddTagging.cs
@@ -1,0 +1,5 @@
+﻿namespace TrafficCourts.Coms.Client.Test.ClientTests;
+
+public class AddTagging : ObjectManagementBase
+{
+}
