+++ src/backend/TrafficCourts/TrafficCourts.Coms.Client.Test/ClientTests/CreateObjects.cs
@@ -3,7 +3,7 @@ using RichardSzalay.MockHttp;
 using System.Net;
 using System.Text.Json;
 
-namespace TrafficCourts.Coms.Client.Test.ComService;
+namespace TrafficCourts.Coms.Client.Test.ClientTests;
 
 public class CreateObjects : ObjectManagementBase
 {
