+++ src/backend/TrafficCourts/TrafficCourts.Coms.Client.Test/ClientTests/DeleteMetadata.cs
@@ -1,0 +1,5 @@
+﻿namespace TrafficCourts.Coms.Client.Test.ClientTests;
+
+public class DeleteMetadata : ObjectManagementBase
+{
+}
