+++ src/backend/TrafficCourts/TrafficCourts.Coms.Client.Test/ClientTests/ListObjectVersion.cs
@@ -1,4 +1,4 @@
-﻿namespace TrafficCourts.Coms.Client.Test.ComService;
+﻿namespace TrafficCourts.Coms.Client.Test.ClientTests;
 
 public class ListObjectVersion : ObjectManagementBase
 {
