+++ src/backend/TrafficCourts/TrafficCourts.Coms.Client.Test/ClientTests/MockHttpExtensions.cs
@@ -1,8 +1,6 @@
 using RichardSzalay.MockHttp;
-using System.Reflection.PortableExecutable;
-using TrafficCourts.Coms.Client.Test.ComService;
 
-namespace TrafficCourts.Coms.Client.Test.ComService;
+namespace TrafficCourts.Coms.Client.Test.ClientTests;
 
 public static class MockHttpExtensions
 {
