+++ src/backend/TrafficCourts/TrafficCourts.Coms.Client.Test/ClientTests/RemovePermissions.cs
@@ -1,4 +1,4 @@
-﻿namespace TrafficCourts.Coms.Client.Test.ComService;
+﻿namespace TrafficCourts.Coms.Client.Test.ClientTests;
 
 public class RemovePermissions : ObjectManagementBase
 {
