--- src/backend/TrafficCourts/TrafficCourts.Coms.Client.Test/ComService/AddMetadata.cs
@@ -1,5 +1,0 @@
-﻿namespace TrafficCourts.Coms.Client.Test.ComService;
-
-public class AddMetadata : ObjectManagementBase
-{
-}
