--- src/backend/TrafficCourts/TrafficCourts.Coms.Client.Test/ComService/DeleteTagging.cs
@@ -1,5 +1,0 @@
-﻿namespace TrafficCourts.Coms.Client.Test.ComService;
-
-public class DeleteTagging : ObjectManagementBase
-{
-}
