--- src/backend/TrafficCourts/TrafficCourts.Coms.Client.Test/ComService/ReplaceTagging.cs
@@ -1,5 +1,0 @@
-﻿namespace TrafficCourts.Coms.Client.Test.ComService;
-
-public class ReplaceTagging : ObjectManagementBase
-{
-}
