--- src/backend/TrafficCourts/TrafficCourts.Coms.Client.Test/ComService/SearchUsers.cs
@@ -1,5 +1,0 @@
-﻿namespace TrafficCourts.Coms.Client.Test.ComService;
-
-public class SearchUsers : ObjectManagementBase
-{
-}
