+++ src/backend/TrafficCourts/TrafficCourts.Coms.Client.Test/ServiceTests/CreateFileAsync.cs
@@ -1,6 +1,6 @@
 ﻿using Moq;
 
-namespace TrafficCourts.Coms.Client.Test.ObjectManagementServiceTests;
+namespace TrafficCourts.Coms.Client.Test.ServiceTests;
 
 public class CreateFileAsync : ObjectManagementServiceTest
 {
