+++ src/backend/TrafficCourts/TrafficCourts.Coms.Client.Test/ServiceTests/FileSearchAsync.cs
@@ -1,0 +1,5 @@
+﻿namespace TrafficCourts.Coms.Client.Test.ServiceTests;
+
+public class FileSearchAsync : ObjectManagementServiceTest
+{
+}
