+++ src/backend/TrafficCourts/TrafficCourts.Coms.Client.Test/ServiceTests/UpdateFileAsync.cs
@@ -1,0 +1,5 @@
+﻿namespace TrafficCourts.Coms.Client.Test.ServiceTests;
+
+public class UpdateFileAsync : ObjectManagementServiceTest
+{
+}
